+++ eosq/src/__mocks__/localstorage.mock.ts
@@ -1,5 +1,4 @@
 export class MockStorage implements Storage {
-  [name: string]: any
   length = 0
 
   clearMock = jest.fn()