+++ eosq/src/models/config.ts
@@ -13,58 +13,38 @@ const isEnvSet = (value: string | undefined): boolean => value != null && value 
 
 const newDefaultConfig = () => {
   const core = {
+    version: 1,
     current_network: process.env.REACT_APP_EOSQ_CURRENT_NETWORK || "local",
+    chain_core_symbol: "4,EOS",
     dfuse_auth_endpoint: process.env.REACT_APP_DFUSE_AUTH_URL || "null://",
     dfuse_io_api_key: process.env.REACT_APP_DFUSE_API_KEY || "web_1234567890abc",
     dfuse_io_endpoint: process.env.REACT_APP_DFUSE_API_NETWORK || "localhost:8080",
     secure: process.env.REACT_APP_DFUSE_API_NETWORK_SECURE === "true",
     display_price: false,
-    price_ticker_name: "EOS",
-    version: 1,
     available_networks: [
       {
         id: "local",
         is_test: true,
-        logo: "/images/eos-mainnet.png",
         name: "Local Network",
         url: "http://localhost:8080"
       },
       {
-        id: "eos-mainnet",
-        is_test: false,
-        logo: "/images/eos-mainnet.png",
-        name: "EOS Mainnet",
-        url: "https://eosq.app"
-      },
-      {
         id: "eos-kylin",
         is_test: true,
-        logo: "/images/eos-kylin.png",
         name: "Kylin Testnet",
         url: "https://kylin.eosq.app"
       },
       {
-        id: "eos-eosio",
-        is_test: true,
-        logo: "/images/eos-eosio.png",
-        name: "EOSIO Testnet",
-        url: "https://eosio.eosq.app"
-      },
-      {
-        id: "eos-worbli",
-        is_test: false,
-        logo: "/images/eos-worbli.png",
-        name: "Worbli",
-        url: "https://worbli.eosq.app"
-      },
-      {
         id: "wax-mainnet",
-        is_test: true,
-        logo: "/images/wax-mainnet.png",
+        is_test: false,
         name: "WAX Mainnet",
         url: "https://wax.eosq.app"
       }
     ]
+  }
+
+  if (isEnvSet(process.env.REACT_APP_EOSQ_CHAIN_CORE_SYMBOL)) {
+    core.chain_core_symbol = process.env.REACT_APP_EOSQ_CHAIN_CORE_SYMBOL!
   }
 
   if (isEnvSet(process.env.REACT_APP_EOSQ_DISPLAY_PRICE)) {
@@ -89,9 +69,9 @@ if (!windowTS.TopLevelConfig) {
 export interface EosqNetwork {
   id: string
   name: string
-  is_test: false
-  logo: string
   url: string
+  is_test?: boolean
+  logo?: string
 }
 
 interface EosqConfig {
@@ -101,15 +81,27 @@ interface EosqConfig {
   dfuse_io_endpoint: string
   dfuse_io_api_key: string
   dfuse_auth_endpoint: string
-
-  current_network: string
-  available_networks: EosqNetwork[]
-  display_price: boolean
-  price_ticker_name: string
-
   secure: boolean
+
+  available_networks: EosqNetwork[]
+  current_network: string
+
+  chain_core_symbol: string
+  chain_core_symbol_code: string
+  chain_core_symbol_precision: number
+
+  display_price: boolean
   disable_segments: boolean
   disable_sentry: boolean
 }
 
-export const Config = { ...windowTS.TopLevelConfig, isLocalhost } as EosqConfig
+export const Config = {
+  ...windowTS.TopLevelConfig,
+  chain_core_symbol_precision: parseInt(windowTS.TopLevelConfig.chain_core_symbol.split(",")[0]),
+  chain_core_symbol_code: windowTS.TopLevelConfig.chain_core_symbol.split(",")[1],
+  isLocalhost
+} as EosqConfig
+
+export const getActiveNetworkConfig = (): EosqNetwork | undefined => {
+  return Config.available_networks.find((network) => network.id === Config.current_network)
+}