+++ eosq/src/pages/account/votes/account-votes.tsx
@@ -15,6 +15,7 @@ import { calculateVoteStrength } from "./vote.helpers"
 import { Stream } from "@dfuse/client"
 import { registerAccountDetailsListeners } from "../../../streams/account-listeners"
 import { metricsStore } from "../../../stores"
+import { Config } from '../../../models/config'
 
 interface Props {
   account: Account
@@ -195,14 +196,14 @@ export class AccountVotes extends React.Component<Props, State> {
           variant="compact"
           label={t("account.summary.voter_info.labels.vote_weight")}
         >
-          {numeral(voteWeight).format("0,0.0000")} EOS
+          {numeral(voteWeight).format("0,0.0000")} {Config.chain_core_symbol_code}
         </DetailLine>
         <DetailLine
           mb={2}
           variant="compact"
           label={t("account.summary.voter_info.labels.decayed_vote_weight")}
         >
-          {numeral(decayedVoteWeight).format("0,0.0000")} EOS
+          {numeral(decayedVoteWeight).format("0,0.0000")} {Config.chain_core_symbol_code}
         </DetailLine>
       </Cell>
     )