+++ eosq/src/pages/vote/vote-tally.page.tsx
@@ -7,9 +7,6 @@ import { Vote } from "../../models/vote"
 import { metricsStore, voteStore } from "../../stores"
 import { PageContainer } from "../../components/page-container/page-container"
 import { registerVoteTallyStream } from "../../streams/vote-listener"
-
-// temp ignore for dev
-
 import { DataError, Box } from "@dfuse/explorer"
 import { Stream, ErrorData } from "@dfuse/client"