+++ eosq/src/services/block.ts
@@ -1,7 +1,7 @@
 import { task } from "mobx-task"
 import { BlockSummary } from "../models/block"
 import { blockStore } from "../stores"
-import { log } from "./logger"
+import { debugLog } from "./logger"
 import {
   getBlock as getBlockApi,
   listBlocks,
@@ -11,20 +11,20 @@ import {
 
 export const fetchBlock = task(
   async (blockId: string) => {
-    log.info("Searching for block id [%s].", blockId)
+    debugLog("Searching for block id [%s].", blockId)
     const block = blockStore.findById(blockId)
     if (block !== undefined) {
-      log.info("Found block [%s] in blocks cache.", blockId, block)
+      debugLog("Found block [%s] in blocks cache.", blockId, block)
       return block
     }
 
     const result = await getBlock(blockId)
     if (result !== null) {
-      log.info("Found block [%s] via search API.", blockId, result)
+      debugLog("Found block [%s] via search API.", blockId, result)
       return result as BlockSummary
     }
 
-    log.info("Block [%s] not found anywhere.", blockId)
+    debugLog("Block [%s] not found anywhere.", blockId)
     return null
   },
   { swallow: true }
@@ -42,7 +42,7 @@ export const getBlocks = task(
   async (offset: number, perPage: number) => {
     const blocks = await listBlocks(offset, perPage)
     if (!blocks || blocks.length === 0) {
-      log.info("No account found for query [%s] via API.")
+      debugLog("No account found for query [%s] via API.")
       return null
     }
 
@@ -55,7 +55,7 @@ export const getBlock = task(
   async (id: string) => {
     const block = await getBlockApi(id)
     if (!block) {
-      log.info("No block found for query [%s] via API.")
+      debugLog("No block found for query [%s] via API.")
       return null
     }
 
@@ -71,7 +71,7 @@ function isEmptyTransactionResponse(response?: ListTransactionsResponse): boolea
 export const fetchBlockTransactions = async (id: string, cursor: string, perPage: number) => {
   const transactionResponse = await listBlockTransactions(id, cursor, perPage)
   if (isEmptyTransactionResponse(transactionResponse)) {
-    log.info("No transactions found for block [%s] via API.", id)
+    debugLog("No transactions found for block [%s] via API.", id)
     return null
   }