+++ eosq/src/setupTests.ts
@@ -1,3 +1,5 @@
+import { createDfuseClient } from '@dfuse/client'
+import { initializeDfuseClient } from '@dfuse/explorer'
 import { configure } from "enzyme"
 import Adapter from "enzyme-adapter-react-16"
 
@@ -5,6 +7,11 @@ import "jest-enzyme"
 import "jest-localstorage-mock"
 
 configure({ adapter: new Adapter() })
+initializeDfuseClient(createDfuseClient({
+  apiKey: "web_123456789",
+  network: "localhost",
+  requestIdGenerator: jest.fn(() => "dc-123"),
+}))
 
 // Initialize correct i18n resources
 withConsoleDisabled(() => {